Start adding some Spanish strings
An issue here is the triple-pane menu to configure settings on the
modules, as given the text "Module Settings", "Popup settings", etc.
is "Configuración de módulo" in Spanish, but that is exponentially
longer and beings to get truncated on the View. Need to figure out
a better wording for that area.